requirements.txt
discord.py[voice]
youtube-dlc
